+++ src/utils/reader/index.ts
@@ -7,3 +7,4 @@ export * from './speech';
 export * from './detect';
 export * from './contrast';
 export * from './palette';
+export * from './wordHighlight';